fix(conformance): stop /StructParents keys wrapping into each other's key space
A page's /StructParents and the /ParentTree /Nums keys were both narrowed to int without a range
check, into one shared key space. So /StructParents 4294967296 became 0, and the page was handed
a different page's MCID-to-element map — inheriting its /Lang, its element types and its alt text.

The consequence is the one this rule set exists to prevent. A two-page document where page A
carries /StructParents 0 with a ParentTree entry supplying /Lang, and page B carries
/StructParents 4294967296 over text with no language anywhere, is failed by veraPDF on
ISO14289-1:7.2-34 and was reported PDF/UA-1 clean here: page B resolved page A's map, found a
/Lang in the hierarchy and skipped the check. A preflight tool certifying a document the
reference implementation rejects is worse than one that declines to judge it.

Out of range is now treated as "no such key" rather than narrowed, which is both the FP-safe
direction this walker takes everywhere else and what a conformant reader sees — no entry with
that key exists.

The same value is read by five annotation rules through StructParentOf, each narrowing it
separately. Rather than five call-site checks, StructParentOf gains a long-taking overload that
range-checks once, and the callers pass PdfInteger.Value through unnarrowed. That is the shape
that survives someone adding a sixth caller.

Conformance suite unchanged against the oracle: 1,237 passed with veraPDF available, 0 skipped.

+++ src/VellumPdf.Conformance/Rules/Ua/UaFormFieldAltRule.cs
@@ -253,7 +253,7 @@ internal sealed class UaFormFieldAltRule : IConformanceRule
     {
         if (context.Resolve(widget.Get(_structParent)) is not PdfInteger spInt)
             return false;
-        var node = tree.StructParentOf((int)spInt.Value);
+        var node = tree.StructParentOf(spInt.Value);
         if (node is null)
             return false;
         return node.Dict.Get(_alt) is not null;
